requirements.txt
# RAG Using Azure Open AI 

openai
python-dotenv
pandas
numpy
python-docx
azure
azure-ai-documentintelligence
openpyxl

# Listed all the requirements here - This is for chatbot

aiofiles==23.2.1
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==24.2.0
bidict==0.23.1
cachetools==5.4.0
certifi==2024.7.4
chainlit==1.1.402
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.110.3
filelock==3.15.4
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.6.1
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-speech==2.27.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
groq==0.9.0
grpcio==1.65.5
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==8.0.0
jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-groq==0.1.9
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.2
langsmith==0.1.99
Lazify==0.4.0
literalai==0.0.607
MarkupSafe==2.1.5
marshmallow==3.21.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.84
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
orjson==3.10.7
packaging==23.2
pandas==2.2.2
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
proto-plus==1.24.0
protobuf==4.25.4
py-dotenv==0.1
pyasn1==0.6.0
pyasn1-modules==0.4.0
PyAudio==0.2.14
pydantic==2.8.2
pydantic-core==2.20.1
PyJWT==2.9.0
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
sympy==1.13.2
syncer==2.0.3
tenacity==8.5.0
tomli==2.0.1
torch==2.4.0
tqdm==4.66.5
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
uptrace==1.26.0
urllib3==2.2.2
uvicorn==0.25.0
watchfiles==0.20.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.20.0


code

# Since Azure Open AI models are not giving the proper response. Developed a chatbot by using Chainlit framework, I'm utilizing other open source models.

from langchain_community.document_loaders import PyPDFLoader # For loading the Pdf
from langchain.prompts import PromptTemplate # Prompt template
from langchain_pinecone import PineconeVectorStore # Vector Database
from langchain.text_splitter import RecursiveCharacterTextSplitter # For Chunking
from langchain.chains import RetrievalQA # For Retrieval
from langchain_groq import ChatGroq # Inference Engine
from dotenv import load_dotenv # For detecting env variables
from langchain.embeddings import OllamaEmbeddings # To perform vector embeddings
import chainlit as cl # For user interface
from langchain_groq import ChatGroq # Inference Engine

load_dotenv() # Detecting env 

# Defining prompt 
prompt_template = """ 

    You are a financial analyst expert at summarizing the complex amazon financial report that has been given to you.
    
    Analyze the amazon financial report and provide
        1. A detailed performance summary
        2. Calculate the following financial ratios
        3. Summarize the key risk factors
        
           Output Format:
        - Performance Summary: [Comprehensive overview of financial performance]
        - Financial Ratios (Table):
            Current Ratio
            Debt-to-Equity Ratio
            Return on Equity (ROE)
            Return on Assets (ROA)
            Gross Profit Margin
            Net Profit Margin
            Earnings Per Share (EPS)
        - Risk Factors Summary: [Key risks identified in the report]

    Context: {context} Question: {question}

    Helpful answer:
    
"""

# Just created a function to interact with the prompt template
def set_custom_prompt():
    prompt = PromptTemplate(template=prompt_template, input_variables=['context', 'question'])  
    return prompt

# Defined this function to perform retrieval
def retrieval_qa_chain(llm, prompt, db):
    qa_chain = RetrievalQA.from_chain_type(
        llm, retriever=db.as_retriever(), chain_type_kwargs={"prompt": prompt}
    )
    return qa_chain

# This function is for defining llm model
def load_llm():
    groqllm = ChatGroq(
        model ="llama3-8b-8192", temperature=0
    )
    return groqllm

# Here just loading the pdf and transforming it to chunks, and performing vector embeddings as well as storing the vector embeddings in Pinecone vector database.
def qa_bot():
    
    data = PyPDFLoader('C:\\Users\\USER\\OneDrive\\Desktop\\NeoStats Hackathon\\Amazon-com-Inc-2023-Annual-Report.pdf')
    loader = data.load()
    chunk = RecursiveCharacterTextSplitter(chunk_size=8000, chunk_overlap=0)
    splitdocs = chunk.split_documents(loader)
    index_name = "langchain2"
    db = PineconeVectorStore.from_documents(splitdocs[:5], OllamaEmbeddings(model ="mxbai-embed-large"),index_name=index_name)
    llm = load_llm()
    qa_prompt = set_custom_prompt()
    qa = retrieval_qa_chain(llm, qa_prompt, db)
    return qa


# Chainlit decorator for starting the app
@cl.on_chat_start
async def start():
    chain = qa_bot()
    msg = cl.Message(content="Starting the bot...")
    await msg.send()
    msg.content = "Hi, Welcome to the Financial Analyzer Bot. Ask questions about amazon annual report"
    await msg.update()

    cl.user_session.set("chain", chain)

@cl.on_message
async def main(message: cl.Message):
    chain = cl.user_session.get("chain")
    if chain is None:
        return
    try:
        res = await chain.acall({'query': message.content})
        answer = res['result']
        await cl.Message(content=answer).send()
    except Exception as e:
        await cl.Message(content=f"An error occurred: {e}").send()



